# docs/assets/openapi.yaml
openapi: 3.1.0
info:
  title: DeskBot API
  version: 0.1.0
  summary: REST API for monitoring and controlling DeskBot.
  description: |
    HTTP API for monitoring and controlling a DeskBot instance.

    The API exposes robot state, configuration, perception, audio and
    conversation status, conversational history, speech and emotion commands,
    hardware calibration, user preferences, learning, and performance
    profiling endpoints.

    The WebSocket event stream is documented with the `x-websocket` extension
    because standard OpenAPI does not define WebSocket operations.
servers:
  - url: http://localhost:8000/api/v1
    description: Local DeskBot API
...
paths:
...
  /state:
    get:
      tags: [state]
      summary: Get current robot state
      operationId: getState
      responses:
        "200":
          description: Current robot state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StateResponse"
              examples:
                attached:
                  value:
                    state: idle
                unavailable:
                  value:
                    state: unknown
                    detail: DeskBot app not attached

    post:
      tags: [commands]
      summary: Transition robot state
      operationId: setState
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StateRequest"
            example:
              state: curious
      responses:
        "200":
          description: Command accepted or rejected by the running application.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"
              examples:
                success:
                  value:
                    status: ok
                    state: curious
                invalid:
                  value:
                    status: error
                    detail: "Invalid state. Valid: ['boot', 'idle', 'curious', 'listening', 'thinking', 'speaking', 'sleeping']"
                unavailable:
                  value:
                    status: error
                    detail: DeskBot app not attached
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /speak:
    post:
      tags: [commands]
      summary: Send text through the conversation pipeline
      description: |
        Publishes the supplied text as a `SpeechRecognized` event so the
        conversation pipeline can process it through the configured LLM/TTS
        flow.

        This is different from `/speak-direct`, which bypasses STT and LLM.
      operationId: speak
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakRequest"
            example:
              text: Hello DeskBot
      responses:
        "200":
          description: Command accepted or rejected by the running application.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"
              examples:
                success:
                  value:
                    status: ok
                    text: Hello DeskBot
                unavailable:
                  value:
                    status: error
                    detail: DeskBot app not attached
        "422":
          $ref: "#/components/responses/ValidationError"
  /speak-direct:
    post:
      tags: [commands]
      summary: Speak text directly through TTS
      description: |
        Synthesizes and plays the supplied text directly through the configured
        TTS provider. The STT and LLM stages are bypassed.
      operationId: speakDirect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakRequest"
            example:
              text: Direct TTS test
      responses:
        "200":
          description: Command accepted or rejected by the running application.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"
              examples:
                success:
                  value:
                    status: ok
                    text: Direct TTS test
                unavailable:
                  value:
                    status: error
                    detail: DeskBot or TTS not attached
        "422":
          $ref: "#/components/responses/ValidationError"
  /emotion:
    post:
      tags: [commands]
      summary: Set current emotion
      description: |
        Publishes an emotion change event.

        The API currently accepts the emotion as a string and returns a
        command-level error for an unknown value rather than using an HTTP 4xx
        response.
      operationId: setEmotion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmotionRequest"
            examples:
              happy:
                value:
                  emotion: happy
                  intensity: 0.9
              curious:
                value:
                  emotion: curious
                  intensity: 1.0
      responses:
        "200":
          description: Command accepted or rejected by the running application.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"
              examples:
                success:
                  value:
                    status: ok
                    emotion: happy
                invalid:
                  value:
                    status: error
                    detail: "Invalid emotion. Valid: ['neutral', 'happy', 'curious', 'thinking', 'sleepy', 'embarrassed', 'excited', 'sad', 'surprised', 'angry']"
                unavailable:
                  value:
                    status: error
                    detail: DeskBot app not attached
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /calibration/servos:
    get:
      tags: [calibration]
      summary: List servos
      operationId: listServos
      responses:
        "200":
          description: Current servo positions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServoListResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /calibration/servos/{name}/move:
    post:
      tags: [calibration]
      summary: Move a servo to an angle
      operationId: moveServo
      parameters:
        - $ref: "#/components/parameters/ServoName"
        - name: angle
          in: query
          required: true
          description: Target servo angle in degrees.
          schema:
            type: number
            format: float
          example: 90
        - name: duration_s
          in: query
          required: false
          description: Movement duration in seconds.
          schema:
            type: number
            format: float
            default: 0.4
            minimum: 0
          example: 0.4
      responses:
        "200":
          description: Servo movement completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServoMoveResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /calibration/servos/{name}/release:
    post:
      tags: [calibration]
      summary: Release a servo
      description: Stops actively driving the named servo so it can be moved by hand.
      operationId: releaseServo
      parameters:
        - $ref: "#/components/parameters/ServoName"
      responses:
        "200":
          description: Servo released.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServoReleaseResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /calibration/servos/release_all:
    post:
      tags: [calibration]
      summary: Release all servos
      operationId: releaseAllServos
      responses:
        "200":
          description: All servos released.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseAllResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /calibration/servos/calibrate/{name}:
    post:
      tags: [calibration]
      summary: Run servo calibration sequence
      description: |
        Moves the selected servo through minimum, centre, maximum, and centre
        positions with pauses between movements.
      operationId: calibrateServo
      parameters:
        - $ref: "#/components/parameters/ServoName"
      responses:
        "200":
          description: Calibration sequence completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServoCalibrationResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /calibration/display:
    get:
      tags: [calibration]
      summary: Get display configuration
      operationId: getDisplayConfig
      responses:
        "200":
          description: Current display configuration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisplayConfig"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /calibration/display/test_pattern:
    post:
      tags: [calibration]
      summary: Show a display test pattern
      operationId: showDisplayTestPattern
      parameters:
        - name: pattern
          in: query
          required: false
          description: Test pattern to display.
          schema:
            type: string
            enum: [gradient, grid, solid, checkerboard]
            default: gradient
          example: grid
      responses:
        "200":
          description: Test pattern displayed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisplayTestPatternResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /calibration/display/clear:
    post:
      tags: [calibration]
      summary: Clear the display
      operationId: clearDisplay
      responses:
        "200":
          description: Display cleared.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisplayClearResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  parameters:
...
    ServoName:
      name: name
      in: path
      required: true
      description: Configured servo name.
      schema:
        type: string
      example: head_pan
...
  responses:
    BadRequest:
      description: Request was invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
          example:
            detail: "Unknown pattern: 'foo'"
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
          example:
            detail: Conversation not found
    ServiceUnavailable:
      description: Required DeskBot service or hardware is not available.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
          example:
            detail: No servo controller available
    ValidationError:
      description: Request body failed FastAPI/Pydantic validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPValidationError"
  schemas:
...
    StateRequest:
      type: object
      required: [state]
      properties:
        state:
          type: string
          description: |
            Target robot state. The server currently validates this value
            inside the command handler rather than through the request schema.
          example: idle
    StateResponse:
      type: object
      required: [state]
      properties:
        state:
          type: string
          example: idle
        detail:
          type: string
          description: Present when the DeskBot application is not attached.
    SpeakRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 1000
          description: Text to process or speak.
          example: Hello DeskBot
    EmotionRequest:
      type: object
      required: [emotion]
      properties:
        emotion:
          type: string
          description: |
            Emotion name. Supported values are neutral, happy, curious,
            thinking, sleepy, embarrassed, excited, sad, surprised, and angry.
          example: happy
        intensity:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 1
          description: Emotion intensity.
    CommandResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok, error]
        detail:
          type: string
          description: Present when the command was rejected.
        text:
          type: string
          description: Echo of the requested speech text.
        emotion:
          type: string
          description: Resulting emotion.
        state:
          type: string
          description: Resulting robot state.
...
    Servo:
      type: object
      required: [name, angle]
      properties:
        name:
          type: string
        angle:
          type: number
          format: float
    ServoListResponse:
      type: object
      required: [servos]
      properties:
        servos:
          type: array
          items:
            $ref: "#/components/schemas/Servo"
    ServoMoveResponse:
      type: object
      required: [name, angle, duration_s]
      properties:
        name:
          type: string
        angle:
          type: number
          format: float
        duration_s:
          type: number
          format: float
    ServoReleaseResponse:
      type: object
      required: [name, released]
      properties:
        name:
          type: string
        released:
          type: boolean
    ReleaseAllResponse:
      type: object
      required: [released]
      properties:
        released:
          type: boolean
    ServoCalibrationStep:
      type: object
      required: [position, angle, actual_angle]
      properties:
        position:
          type: string
          enum: [min, centre, max]
        angle:
          type: number
          format: float
        actual_angle:
          type: number
          format: float
    ServoCalibrationResponse:
      type: object
      required: [servo, sequence]
      properties:
        servo:
          type: string
        sequence:
          type: array
          items:
            $ref: "#/components/schemas/ServoCalibrationStep"
    DisplayConfig:
      type: object
      required: [backend, width, height, rotation, spi_hz, invert]
      properties:
        backend:
          type: string
        width:
          type: integer
        height:
          type: integer
        rotation:
          type: integer
        spi_hz:
          type: integer
        invert:
          type: boolean
    DisplayTestPatternResponse:
      type: object
      required: [pattern, width, height]
      properties:
        pattern:
          type: string
          enum: [gradient, grid, solid, checkerboard]
        width:
          type: integer
        height:
          type: integer
    DisplayClearResponse:
      type: object
      required: [cleared]
      properties:
        cleared:
          type: boolean
    HTTPError:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
        input: {}
        ctx:
          type: object
          additionalProperties: true
    HTTPValidationError:
      type: object
      required: [detail]
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
...